Rama actualizada

+++ backend/src/main/java/com/co/service/BusService.java
@@ -1,6 +1,10 @@
 package com.co.service;
 
+import com.co.conversion.BusDTOConverter;
+import com.co.dto.BusDTO;
+import com.co.dto.ConductorDTO;
 import com.co.model.Bus;
+import com.co.model.Conductor;
 import com.co.repository.BusRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -14,27 +18,41 @@ public class BusService {
     @Autowired
     private BusRepository busRepository;
 
+    @Autowired
+    private BusDTOConverter busDTOConverter;
+
     // Obtener todos los buses
-    public List<Bus> findAll() {
+    public List<Bus> getAllBuses() {
         return busRepository.findAll();
     }
 
     // Obtener un bus por ID
-    public Optional<Bus> findById(Long id) {
-        return busRepository.findById(id);
+    public BusDTO getBus(Long id) {
+        return busDTOConverter.entityToDTO(busRepository.findById(id).orElseThrow());
     }
+
 
     // Crear o actualizar un bus
-    public void save(Bus bus) {
-        busRepository.save(bus);
-    }
-
-    public List<Bus> findByIds(List<Long> ids) {
-        return busRepository.findByIdIn(ids);
+    public BusDTO save(BusDTO busDTO) {
+        Bus bus = busDTOConverter.DTOToEntity(busDTO);
+        return busDTOConverter.entityToDTO(busRepository.save(bus));
     }
 
     // Eliminar un bus por ID
     public void delete(Long id) {
+        if (!busRepository.existsById(id)) {
+            throw new RuntimeException("Bus no encontrado para eliminar");
+        }
         busRepository.deleteById(id);
     }
+
+    public BusDTO createBus(BusDTO busDTO) {
+        Bus bus = busDTOConverter.DTOToEntity(busDTO);
+        return busDTOConverter.entityToDTO(busRepository.save(bus));
+    }
+
+    public List<Bus> findByIds(List<Long> ids) {
+        return busRepository.findAllById(ids);
+    }
 }
+
